Lib

Documentation

fibN :: (Num a, Enum a, Ord a) => a -> [a]

Calculate a list of Fibonacci numbers of length len.

fibBetter :: Integer -> [Integer]

Better version, recursively construct a list.

fibZip :: Int -> [Integer]

Short version, using zip with the same list twice.

goldenRatio :: Int -> Double

Calculate the golden ratio.